requirements.txt
openai==2.9.0
python-dotenv==1.2.1
requests==2.32.5
